requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
Django==3.1
django-bootstrap4==2.2.0
django-cleanup==5.0.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
easy-thumbnails==2.7
importlib-metadata==1.7.0
Pillow==7.2.0
psycopg2-binary==2.8.5
pytz==2020.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
zipp==3.1.0
